requirements.txt
# Core ML and Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter and Notebook Support
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.7.0
